requirements.txt
flask
nltk
pdfplumber
python-docx
scikit-learn
